+++ spec/rubocop/cop/rspec/rails/http_status_spec.rb
@@ -83,4 +83,69 @@ RSpec.describe RuboCop::Cop::RSpec::Rails::HttpStatus do
       end
     end
   end
+
+  context 'when EnforcedStyle is `be_status`' do
+    let(:cop_config) { { 'EnforcedStyle' => 'be_status' } }
+
+    it 'registers an offense when using numeric value' do
+      expect_offense(<<-RUBY)
+        it { is_expected.to have_http_status 200 }
+                            ^^^^^^^^^^^^^^^^^^^^ Prefer `be_ok` over `have_http_status 200` to describe HTTP status code.
+      RUBY
+
+      expect_correction(<<-RUBY)
+        it { is_expected.to be_ok }
+      RUBY
+    end
+
+    it 'registers an offense when using symbolic value' do
+      expect_offense(<<-RUBY)
+        it { is_expected.to have_http_status :ok }
+                            ^^^^^^^^^^^^^^^^^^^^ Prefer `be_ok` over `have_http_status :ok` to describe HTTP status code.
+      RUBY
+
+      expect_correction(<<-RUBY)
+        it { is_expected.to be_ok }
+      RUBY
+    end
+
+    it 'does not register an offense when using numeric value' do
+      expect_no_offenses(<<-RUBY)
+        it { is_expected.to be_ok }
+      RUBY
+    end
+
+    it 'does not register an offense when using allowed symbols' do
+      expect_no_offenses(<<-RUBY)
+        it { is_expected.to have_http_status :error }
+        it { is_expected.to have_http_status :success }
+        it { is_expected.to have_http_status :missing }
+        it { is_expected.to have_http_status :redirect }
+      RUBY
+    end
+
+    context 'with parenthesis' do
+      it 'registers an offense when using numeric value' do
+        expect_offense(<<-RUBY)
+          it { is_expected.to have_http_status(404) }
+                              ^^^^^^^^^^^^^^^^^^^^^ Prefer `be_not_found` over `have_http_status(404)` to describe HTTP status code.
+        RUBY
+
+        expect_correction(<<-RUBY)
+          it { is_expected.to be_not_found }
+        RUBY
+      end
+
+      it 'registers an offense when using symbolic value' do
+        expect_offense(<<-RUBY)
+          it { is_expected.to have_http_status(:not_found) }
+                              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer `be_not_found` over `have_http_status(:not_found)` to describe HTTP status code.
+        RUBY
+
+        expect_correction(<<-RUBY)
+          it { is_expected.to be_not_found }
+        RUBY
+      end
+    end
+  end
 end